Arrange packages, classses and Add jwt dependency
