+++ backend/src/main/java/org/nexters/cultureland/api/user/service/impl/FacebookSSOServiceImpl.java
@@ -1,6 +1,6 @@
 package org.nexters.cultureland.api.user.service.impl;
 
-import org.nexters.cultureland.common.FacebookUserResponse;
+import org.nexters.cultureland.api.user.response.FacebookUserResponse;
 import org.nexters.cultureland.api.user.model.User;
 import org.nexters.cultureland.api.user.repo.UserRepository;
 import org.nexters.cultureland.api.user.service.SSOService;
