+++ backend/src/main/java/org/nexters/cultureland/api/user/service/impl/UserServiceImpl.java
@@ -1,6 +1,6 @@
 package org.nexters.cultureland.api.user.service.impl;
 
-import org.nexters.cultureland.api.user.exception.BadRequestException;
+import org.nexters.cultureland.common.excepion.BadRequestException;
 import org.nexters.cultureland.api.user.model.User;
 import org.nexters.cultureland.api.user.repo.UserRepository;
 import org.nexters.cultureland.api.user.service.UserService;
