+++ backend/src/main/java/org/nexters/cultureland/common/excepion/UnauthorizedException.java
@@ -1,4 +1,4 @@
-package org.nexters.cultureland.api.user.exception;
+package org.nexters.cultureland.common.excepion;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
