--- backend/src/main/java/org/nexters/cultureland/config/CommonConfig.java
@@ -1,31 +1,0 @@
-package org.nexters.cultureland.config;
-
-import org.apache.http.client.HttpClient;
-import org.apache.http.impl.client.HttpClientBuilder;
-import org.nexters.cultureland.api.user.handler.RestTemplateErrorHandler;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class CommonConfig {
-    @Bean
-    public RestTemplate restTemplate(){
-        HttpComponentsClientHttpRequestFactory factory = new HttpComponentsClientHttpRequestFactory();
-
-        HttpClient client = HttpClientBuilder.create()
-                .setMaxConnTotal(50)
-                .setMaxConnPerRoute(20)
-                .build();
-
-        factory.setHttpClient(client);
-        factory.setConnectTimeout(2000);
-        factory.setReadTimeout(5000);
-
-        RestTemplate restTemplate = new RestTemplate(factory);
-        restTemplate.setErrorHandler(new RestTemplateErrorHandler());
-        return restTemplate;
-    }
-
-}
